+++ python-package/src/booster.rs
@@ -7,7 +7,7 @@ use perpetual_rs::conformal::cqr::CalData;
 use perpetual_rs::constraints::Constraint;
 use perpetual_rs::data::Matrix;
 use perpetual_rs::objective_functions::Objective;
-use perpetual_rs::UnivariateBooster as CratePerpetualBooster;
+use perpetual_rs::PerpetualBooster as CratePerpetualBooster;
 use pyo3::exceptions::{PyKeyError, PyValueError};
 use pyo3::prelude::*;
 use pyo3::types::IntoPyDict;
