+++ src/conformal/cqr.rs
@@ -1,10 +1,10 @@
-use crate::objective_functions::Objective;
-use crate::{errors::PerpetualError, utils::percentiles, Matrix, UnivariateBooster};
+use crate::objective_functions::objective::Objective;
+use crate::{errors::PerpetualError, utils::percentiles, Matrix, PerpetualBooster};
 use std::collections::HashMap;
 
 pub type CalData<'a> = (Matrix<'a, f64>, &'a [f64], &'a [f64]); // (x_flat_data, rows, cols), y, alpha
 
-impl UnivariateBooster {
+impl PerpetualBooster {
     /// Calibrate models to get prediction intervals
     /// * `alpha` - Alpha list to train calibration models for
     pub fn calibrate(
@@ -19,17 +19,17 @@ impl UnivariateBooster {
 
         for alpha_ in alpha {
             let lower_quantile = Some(alpha_ / 2.0);
-            let mut model_lower = UnivariateBooster::default().set_objective(Objective::QuantileLoss {
+            let mut model_lower = PerpetualBooster::default().set_objective(Objective::QuantileLoss {
                 quantile: lower_quantile,
             });
-            model_lower.fit(&data, &y, sample_weight, group)?;
+            model_lower.fit(data, y, sample_weight, group)?;
 
             let upper_quantile = Some(1.0 - alpha_ / 2.0);
-            let mut model_upper = UnivariateBooster::default().set_objective(Objective::QuantileLoss {
+            let mut model_upper = PerpetualBooster::default().set_objective(Objective::QuantileLoss {
                 quantile: upper_quantile,
             });
 
-            model_upper.fit(&data, &y, sample_weight, group)?;
+            model_upper.fit(data, y, sample_weight, group)?;
 
             let y_cal_pred_lower = model_lower.predict(&x_cal, true);
             let y_cal_pred_upper = model_upper.predict(&x_cal, true);
@@ -37,8 +37,8 @@ impl UnivariateBooster {
             for i in 0..y_cal.len() {
                 scores.push(f64::max(y_cal_pred_lower[i] - y_cal[i], y_cal[i] - y_cal_pred_upper[i]));
             }
-            let perc = (1.0 - (*alpha_ as f64)) * (1.0 + 1.0 * (1.0 / (scores.len() as f64)));
-            let score = percentiles(&scores, &vec![1.0; scores.len()], &vec![perc])[0];
+            let perc = (1.0 - *alpha_) * (1.0 + 1.0 * (1.0 / (scores.len() as f64)));
+            let score = percentiles(&scores, &vec![1.0; scores.len()], &[perc])[0];
             self.cal_models
                 .insert(alpha_.to_string(), [(model_lower, -score), (model_upper, score)]);
         }
@@ -70,9 +70,9 @@ impl UnivariateBooster {
 #[cfg(test)]
 mod tests {
 
-    use crate::objective_functions::Objective;
+    use crate::objective_functions::objective::Objective;
     use crate::Matrix;
-    use crate::UnivariateBooster;
+    use crate::PerpetualBooster;
     use polars::io::SerReader;
     use polars::prelude::{CsvReadOptions, DataType};
     use std::error::Error;
@@ -164,7 +164,7 @@ mod tests {
         let matrix_train = Matrix::new(&data_train, y_train.len(), 8);
         let matrix_test = Matrix::new(&data_test, y_test.len(), 8);
 
-        let mut model = UnivariateBooster::default()
+        let mut model = PerpetualBooster::default()
             .set_objective(Objective::SquaredLoss)
             .set_max_bin(10)
             .set_budget(0.1);
