+++ src/histogram.rs
@@ -59,6 +59,24 @@ impl<'a> FeatureHistogram<'a> {
         Self { data: unsafe { &*ptr } }
     }
 
+    /// Updates the histogram data based on the provided gradients and Hessian values.
+    ///
+    /// # Arguments
+    /// * `feature`: A slice of bin indices for each data element.
+    /// * `sorted_grad`: A slice of gradients corresponding to the data elements in this node.
+    /// * `sorted_hess`: An optional slice of Hessian values corresponding to the data elements.
+    /// * `index`: A slice of original indices for the data elements in this node.
+    ///
+    /// # Safety
+    /// This function is unsafe because it uses `get_unchecked` and `as_mut().unwrap_unchecked()`,
+    /// which bypass Rust's standard safety checks. The caller must ensure the following:
+    ///
+    /// 1. The `feature` slice must be a valid bin index for every element.
+    /// 2. All indices in the `index` slice must be within the bounds of the `feature` slice.
+    /// 3. The `sorted_grad` slice must have the same length as the `index` slice.
+    /// 4. If `sorted_hess` is `Some`, it must also have the same length as the `index` slice.
+    /// 5. The internal `self.data` structure must not be modified externally while this function is running.
+    /// 6. Each element in `self.data` must contain a valid `Some` value that can be mutated.
     pub unsafe fn update(
         &self,
         feature: &[u16], // an array which shows the bin index for each element of the feature, length = whole length of data
@@ -101,6 +119,15 @@ impl<'a> FeatureHistogram<'a> {
         }
     }
 
+    /// Updates the cut-off values for the histogram bins.
+    ///
+    /// This function is unsafe because...
+    ///
+    /// # Safety
+    /// The `cuts` slice must be sorted in ascending order.
+    /// The length of `cuts` must be exactly one less than the number of bins.
+    /// Calling this function with an unsorted slice or an incorrect length
+    /// could lead to incorrect binning logic and potential data corruption.
     pub unsafe fn update_cuts(&self, cuts: &[f64]) {
         let cuts_mod = &cuts[..(cuts.len() - 1)];
         self.data.iter().enumerate().for_each(|(i, b)| {
@@ -158,8 +185,8 @@ pub struct NodeHistogram<'a> {
     pub data: Vec<FeatureHistogram<'a>>,
 }
 
-impl NodeHistogram<'_> {
-    pub fn from_owned(hist: &mut NodeHistogramOwned) -> NodeHistogram {
+impl<'a> NodeHistogram<'a> {
+    pub fn from_owned(hist: &'a mut NodeHistogramOwned) -> NodeHistogram<'a> {
         let histograms = hist
             .data
             .iter_mut()
@@ -288,23 +315,20 @@ pub fn update_histogram(
     unsafe {
         if pool.current_num_threads() > 1 {
             pool.scope(|s| {
-                for i in 0..hist.data.len() {
+                for (i, &col) in col_index.iter().enumerate().take(hist.data.len()) {
                     let h = hist.data.get_unchecked(i);
-                    let feature = data.get_col(col_index[i]);
+                    let feature = data.get_col(col); // Use the value 'col' directly
                     s.spawn(|_| {
-                        h.update(feature, &sorted_grad, sorted_hess.as_deref(), &index[start..stop]);
+                        h.update(feature, sorted_grad, sorted_hess, &index[start..stop]);
                     });
                 }
             });
         } else {
             col_index.iter().enumerate().for_each(|(i, col)| {
                 println!("i: {:?}, col: {:?}", i, col);
-                hist.data.get_unchecked(i).update(
-                    data.get_col(*col),
-                    &sorted_grad,
-                    sorted_hess.as_deref(),
-                    &index[start..stop],
-                );
+                hist.data
+                    .get_unchecked(i)
+                    .update(data.get_col(*col), sorted_grad, sorted_hess, &index[start..stop]);
             });
         }
     }
@@ -316,7 +340,7 @@ mod tests {
     use crate::histogram::{
         update_histogram, FeatureHistogram, FeatureHistogramOwned, NodeHistogram, NodeHistogramOwned,
     };
-    use crate::objective_functions::{Objective, ObjectiveFunction};
+    use crate::objective_functions::objective::{Objective, ObjectiveFunction};
     use crate::Matrix;
     use approx::assert_relative_eq;
     use std::collections::HashSet;
@@ -325,7 +349,7 @@ mod tests {
     #[test]
     fn test_simple_histogram() {
         // instantiate objective function
-        let objective_function = Objective::LogLoss.as_function();
+        let objective_function = Objective::LogLoss;
 
         let nbins = 90;
 
@@ -383,7 +407,7 @@ mod tests {
     #[test]
     fn test_single_histogram() {
         // instantiate objective function
-        let objective_function = Objective::LogLoss.as_function();
+        let objective_function = Objective::LogLoss;
 
         let nbins = 10;
 
@@ -435,7 +459,7 @@ mod tests {
     #[test]
     fn test_histogram_categorical() {
         // instantiate objective function
-        let objective_function = Objective::LogLoss.as_function();
+        let objective_function = Objective::LogLoss;
 
         let file =
             fs::read_to_string("resources/titanic_train_flat.csv").expect("Something went wrong reading the file");
@@ -500,7 +524,7 @@ mod tests {
     #[test]
     fn test_histogram_parallel() {
         // instantiate objective function
-        let objective_function = Objective::LogLoss.as_function();
+        let objective_function = Objective::LogLoss;
 
         let file =
             fs::read_to_string("resources/titanic_train_flat.csv").expect("Something went wrong reading the file");
