+++ src/lib.rs
@@ -12,6 +12,7 @@ pub mod conformal;
 pub mod constants;
 pub mod constraints;
 pub mod data;
+pub mod decision_tree;
 pub mod errors;
 pub mod grower;
 pub mod histogram;
@@ -20,10 +21,9 @@ pub mod objective_functions;
 pub mod prune;
 pub mod sampler;
 pub mod splitter;
-pub mod tree;
 pub mod utils;
 
 // Individual classes, and functions
-pub use booster::multivariate_booster::MultivariateBooster;
-pub use booster::univariate_booster::UnivariateBooster;
+pub use booster::core::PerpetualBooster;
+pub use booster::multi_output::MultiOutputBooster;
 pub use data::Matrix;
