+++ src/objective_functions/adaptive_huber_loss.rs
@@ -1,8 +1,8 @@
 //! Adaptive Huber Loss function
 //!
 //!
-use super::ObjectiveFunction;
-use crate::metrics::Metric;
+
+use crate::{metrics::evaluation::Metric, objective_functions::objective::ObjectiveFunction};
 use serde::{Deserialize, Serialize};
 
 /// Adaptive Huber Loss
@@ -10,6 +10,7 @@ use serde::{Deserialize, Serialize};
 pub struct AdaptiveHuberLoss {
     pub quantile: Option<f64>,
 }
+
 impl ObjectiveFunction for AdaptiveHuberLoss {
     #[inline]
     fn loss(&self, y: &[f64], yhat: &[f64], sample_weight: Option<&[f64]>, _group: Option<&[u64]>) -> Vec<f32> {
